To prevent prestige related commands from being registered with bukkit, pulled prestige related commands in to their own command class and then conditionally register them if prestiges are enabled.
